create a proper client table
