+++ src/table_schemas.py
@@ -61,27 +61,27 @@ class MainTable(Base):
         return "MainTable"
 
 
-class ClientTable(Base):
-    """
-    The main Table
-    """
-    __tablename__ = client_table_suffix
-    entry_id = db.Column(db.Integer, primary_key=True, unique=True, autoincrement=True)
-    date = db.Column(db.Date, nullable=False)
-    client_type = db.Column(db.Enum(ClientType), nullable=False)
-    tasks = db.Column(db.Unicode)
-    notes = db.Column(db.Unicode)
-
-    def __init__(self, **kwargs):
-        try:
-            params = ClientTableParams(kwargs)
-            self.entry_id = params.entry_id
-            self.date = params.date
-            self.client_type = params.client_type
-            self.tasks = params.tasks
-            self.notes = params.notes
-        except KeyError:
-            pass
-
-    def __repr__(self):
-        return f"Client_table_{self.client_id}"
+# class ClientTable(Base):
+#     """
+#     The main Table
+#     """
+#     __tablename__ = client_table_suffix
+#     entry_id = db.Column(db.Integer, primary_key=True, unique=True, autoincrement=True)
+#     date = db.Column(db.Date, nullable=False)
+#     client_type = db.Column(db.Enum(ClientType), nullable=False)
+#     tasks = db.Column(db.Unicode)
+#     notes = db.Column(db.Unicode)
+#
+#     def __init__(self, **kwargs):
+#         try:
+#             params = ClientTableParams(kwargs)
+#             self.entry_id = params.entry_id
+#             self.date = params.date
+#             self.client_type = params.client_type
+#             self.tasks = params.tasks
+#             self.notes = params.notes
+#         except KeyError:
+#             pass
+#
+#     def __repr__(self):
+#         return f"Client_table_{self.client_id}"
